Fix 32-bit for new error code

+++ src/floats.jl
@@ -90,7 +90,7 @@ const SPECIALS = Trie(["nan"=>NaN, "infinity"=>Inf, "inf"=>Inf])
 
 @inline function defaultparser(io::IO, r::Result{T}; decimal::Union{Char, UInt8}=UInt8('.'), kwargs...) where {T <: Union{Float16, Float32, Float64}}
     setfield!(r, 1, missing)
-    setfield!(r, 3, Int(position(io)))
+    setfield!(r, 3, Int64(position(io)))
     b = 0x00
     code = SUCCESS
     eof(io) && (code |= INVALID | EOF; @goto done)
